Proper stat page
Will need to get the route for it, along with getting the proper component for it functioning.

Also created base Deletion button component.

Also adjusted the image alt text for icon, since it said it was the character AVATAR, when it should have said it was the icon, like it should.